+++ app/components/map/mapView.html
@@ -1,5 +1,3 @@
 <div id="map">
     <d3-map data="d3Data" lable="title"></d3-map>
 </div>
-
-{{mapData.map[0].layout}}
